requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-filter==23.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
sqlparse==0.4.3
tzdata==2023.3
